100x speedup for mixture.py
Stores vcf files in memory instead of loading from disk each time.
